improvements

+++ app/plex.py
@@ -54,10 +54,10 @@ def getUsers():
         if not Users.select().where(Users.username == user.title).exists():
             Users.create(email=user.email, username=user.title,
                          token="None", code="None")
-
-    for dbuser in Users.select():
-        if dbuser.username not in [u.title for u in plexusers]:
-            dbuser.delete_instance()
+    if Users.select():
+        for dbuser in Users.select():
+            if dbuser.username not in [u.title for u in plexusers]:
+                dbuser.delete_instance()
     return plexusers
